modular definition of the different basic circuits
modular definition of the different circuits. the results are untested, but this is the basic framework.

the bitstrings are decomposed by each bit marker, then each bit is compared with it's opposite in the other operand, and the rule defined in `*_MAPPING` is applied. the results are then rebound together.